migrator refactoring + minor fixes

+++ contracts/base/periphery/DystRouter01.sol
@@ -56,7 +56,7 @@ contract DystRouter01 {
   /// @dev Calculates the CREATE2 address for a pair without making any external calls.
   function _pairFor(address tokenA, address tokenB, bool stable) internal view returns (address pair) {
     (address token0, address token1) = _sortTokens(tokenA, tokenB);
-    pair = address(uint160(uint256(keccak256(abi.encodePacked(
+    pair = address(uint160(uint(keccak256(abi.encodePacked(
         hex'ff',
         factory,
         keccak256(abi.encodePacked(token0, token1, stable)),
